bbc2cpm: fix failure to work with fully qualified Acorn names.
Prior to this commit, bbc2cpm relied on the CP/M CCP filename to FCB
scanner putting the destination CP/M name into standard CP/M FCB2 but
this in turn relied on the CCP to correctly work out the end of the 1st,
i.e. Acorn filename.  An Acorn name with multiple dot ('.') characters
would throw off the CCP scanner and the 2nd FCB would be blank or
corrupt.

This commit resolves the issue skipping over the Acorn name itself,
after noting the start and terminating it with a CR as before, and then
parsing the CP/M name into FCB1 itself with code borrowed from SJCCP via
http://mdfs.net/Docs/Comp/CPM/ScanName